[REFACTOR] 마이페이지 api 분류 작업
