+++ src/api/mypageAPI.ts
@@ -1,6 +1,6 @@
 import { instance } from "./api";
 
-export const mypageApi = {
+export const mypageAPI = {
   updateNickname: (nickname: string) => {
     return instance.patch(`/users/name`, { nickname });
   },
@@ -11,11 +11,22 @@ export const mypageApi = {
       },
     });
   },
+
   getAlertSetting: async () => {
     const { data } = await instance.get(`/users/preferences`);
     return data;
   },
   updateAlertSetting: async (payload: Partial<preferencesResponse>) => {
     return await instance.patch(`/users/preferences`, payload);
+  },
+
+  getBlockUser: () => {
+    return instance.get(`/users/block`);
+  },
+  blockUser: (block_id: string) => {
+    return instance.post(`/users/block/${block_id}`);
+  },
+  getFollowing: (id: number, targetId: number) => {
+    return instance.post(`/users/neighbors/${id}`, { targetId });
   },
 };
