+++ src/components/mypage/article/neighbors/BannedByMe.tsx
@@ -2,14 +2,14 @@ import React from "react";
 import { useQuery } from "@tanstack/react-query";
 
 import { QUERY_KEYS } from "@/constants/global/querykeys";
-import { authAPI } from "@/api/authAPI";
+import { mypageAPI } from "@/api/mypageAPI";
 
 import MyNeightborProfile from "./profile/MyNeighborProfile";
 
 const BannedByMe = () => {
   const { data } = useQuery({
     queryKey: [QUERY_KEYS.BLOCKUSER],
-    queryFn: async () => await authAPI.getBlockUser(),
+    queryFn: async () => await mypageAPI.getBlockUser(),
   });
   const neighborList: neighborItem[] = data?.data;
