+++ src/tests/msw/myPageHandler.ts
@@ -1,0 +1,44 @@
+import { rest } from "msw";
+
+export const myPageHandler = [
+  rest.patch("/users/name", (req, res, ctx) => {
+    return res(ctx.status(200));
+  }),
+  rest.patch("/users/image", (req, res, ctx) => {
+    return res(ctx.status(200));
+  }),
+  rest.get("/api/users/times", (req, res, ctx) => {
+    const categories: timeOfLearning[] = [
+      {
+        category: "리액트",
+        timespent: 1,
+      },
+      {
+        category: "Next",
+        timespent: 20000,
+      },
+      {
+        category: "Javascript",
+        timespent: 3,
+      },
+    ];
+    return res(ctx.status(200), ctx.json({ categories }));
+  }),
+  rest.get("/mypage/mywritten", (req, res, ctx) => {
+    const categories: timeOfLearning[] = [
+      {
+        category: "리액트",
+        timespent: 1,
+      },
+      {
+        category: "Next",
+        timespent: 20000,
+      },
+      {
+        category: "Javascript",
+        timespent: 3,
+      },
+    ];
+    return res(ctx.status(200), ctx.json({ categories }));
+  }),
+];
